+++ bindings/react/src/index.js
@@ -40,8 +40,6 @@ import SplitterContent from './components/SplitterContent.jsx';
 import SplitterSide from './components/SplitterSide.jsx';
 import Switch from './components/Switch.jsx';
 import Tab from './components/Tab.jsx';
-import TabActive from './components/TabActive.jsx';
-import TabInactive from './components/TabInactive.jsx';
 import Tabbar from './components/Tabbar.jsx';
 import Toast from './components/Toast.jsx';
 import Toolbar from './components/Toolbar.jsx';
@@ -93,8 +91,6 @@ export {
   SplitterSide,
   Switch,
   Tab,
-  TabActive,
-  TabInactive,
   Tabbar,
   Toast,
   Toolbar,