+++ core/src/elements/ons-tabbar/index.js
@@ -19,23 +19,16 @@ import util from '../../ons/util';
 import platform from '../../ons/platform';
 import internal from '../../ons/internal';
 import autoStyle from '../../ons/autostyle';
+import Swiper from '../../ons/internal/swiper';
 import ModifierUtil from '../../ons/internal/modifier-util';
-import AnimatorFactory from '../../ons/internal/animator-factory';
 import BaseElement from '../base/base-element';
-import {TabbarAnimator, TabbarFadeAnimator, TabbarNoneAnimator, TabbarSlideAnimator} from './animator';
 import TabElement from '../ons-tab';
 import contentReady from '../../ons/content-ready';
 
 const scheme = {
   '.tabbar__content': 'tabbar--*__content',
+  '.tabbar__border': 'tabbar--*__border',
   '.tabbar': 'tabbar--*'
-};
-
-const _animatorDict = {
-  'default': TabbarNoneAnimator,
-  'fade': TabbarFadeAnimator,
-  'slide': TabbarSlideAnimator,
-  'none': TabbarNoneAnimator
 };
 
 const rewritables = {
@@ -48,10 +41,8 @@ const rewritables = {
   }
 };
 
-const generateId = (() => {
-  let i = 0;
-  return () => 'ons-tabbar-gen-' + (i++);
-})();
+const nullPage = internal.nullElement;
+const lerp = (x0, x1, t) => (1 - t) * x0 + t * x1;
 
 /**
  * @element ons-tabbar
@@ -61,6 +52,15 @@ const generateId = (() => {
  *   [ja]タブバーをページ下部に表示するためのコンポーネントです。ons-tabと組み合わせて使うことで、ページを管理できます。[/ja]
  * @codepen pGuDL
  * @tutorial vanilla/Reference/tabbar
+ * @modifier material
+ *   [en]A tabbar in Material Design.[/en]
+ *   [ja][/ja]
+ * @modifier autogrow
+ *   [en]Tabs automatically grow depending on their content instead of having a fixed width.[/en]
+ *   [ja][/ja]
+ * @modifier top-border
+ *   [en]Shows a static border-bottom in tabs for iOS top tabbars.[/en]
+ *   [ja][/ja]
  * @guide multiple-page-navigation
  *  [en]Managing multiple pages.[/en]
  *  [ja]Managing multiple pages[/ja]
@@ -153,8 +153,8 @@ export default class TabbarElement extends BaseElement {
    * @type {String}
    * @default none
    * @description
-   *   [en]Animation name. Available values are `"none"`, `"slide"` and `"fade"`. Default is `"none"`.[/en]
-   *   [ja]ページ読み込み時のアニメーションを指定します。"none"、"fade"、"slide"のいずれかを選択できます。デフォルトは"none"です。[/ja]
+   *   [en]If this attribute is set to `"none"` the transitions will not be animated.[/en]
+   *   [ja][/ja]
    */
 
   /**
@@ -175,64 +175,174 @@ export default class TabbarElement extends BaseElement {
    *   [ja]タブバーの位置を指定します。"bottom"もしくは"top"を選択できます。デフォルトは"bottom"です。[/ja]
    */
 
+  /**
+   * @attribute swipeable
+   * @description
+   *   [en]If this attribute is set the tab bar can be scrolled by drag or swipe.[/en]
+   *   [ja]この属性がある時、タブバーをスワイプやドラッグで移動できるようになります。[/ja]
+   */
+
+  /**
+   * @attribute tab-border
+   * @description
+   *   [en]If this attribute is set the tabs show a dynamic bottom border. Only works for iOS since the border is always visible in Material Design.[/en]
+   *   [ja][/ja]
+   */
+
+  /**
+   * @attribute modifier
+   * @type {String}
+   * @description
+   *   [en]The appearance of the tabbar.[/en]
+   *   [ja]タブバーの表現を指定します。[/ja]
+   */
+
   constructor() {
     super();
-
-    this._tabbarId = generateId();
-
-    contentReady(this, () => {
-      this._compile();
-
-      const activeIndex = this.getAttribute('activeIndex');
-
-      const tabbar = this._tabbarElement;
-      if (activeIndex && tabbar.children.length > activeIndex) {
-        tabbar.children[activeIndex].setAttribute('active', 'true');
-      }
-
-      autoStyle.prepare(this);
-      ModifierUtil.initModifier(this, scheme);
-
-      this._animatorFactory = new AnimatorFactory({
-        animators: _animatorDict,
-        baseClass: TabbarAnimator,
-        baseClassName: 'TabbarAnimator',
-        defaultAnimation: this.getAttribute('animation')
-      });
-    });
+    contentReady(this, () => this._compile());
   }
 
   connectedCallback() {
+    if (!this._swiper) {
+      this._swiper = new Swiper({
+        getElement: () => this._contentElement,
+        getInitialIndex: () => this.getAttribute('activeIndex'),
+        getAutoScrollRatio: this._getAutoScrollRatio.bind(this),
+        isAutoScrollable: () => true,
+        preChangeHook: this._onPreChange.bind(this),
+        postChangeHook: this._onPostChange.bind(this),
+        refreshHook: this._onRefresh.bind(this),
+        scrollHook: this._onScroll.bind(this)
+      });
+
+      contentReady(this, () => {
+        this._tabbarBorder = util.findChild(this._tabbarElement, '.tabbar__border');
+        this._swiper.init({ swipeable: this.hasAttribute('swipeable') });
+      });
+    }
+
     contentReady(this, () => this._updatePosition());
   }
 
-  get _contentElement() {
-    return util.findChild(this, '.tabbar__content');
+  disconnectedCallback() {
+    if (this._swiper && this._swiper.initialized) {
+      this._swiper.dispose();
+      this._swiper = null;
+      this._tabbarBorder = null;
+      this._tabsRect = null;
+    }
+  }
+
+  _normalizeEvent(event) {
+    return { ...event, index: event.activeIndex, tabItem: this.tabs[event.activeIndex] };
+  }
+
+  _onPostChange(event) {
+    event = this._normalizeEvent(event);
+    util.triggerElementEvent(this, 'postchange', event);
+    const page = event.tabItem.pageElement;
+    page && page._show();
+  }
+
+  _onPreChange(event) {
+    event = this._normalizeEvent(event);
+
+    let canceled = false;
+    util.triggerElementEvent(this, 'prechange', { ...event, cancel: () => canceled = true });
+
+    if (!canceled) {
+      const { activeIndex, lastActiveIndex } = event;
+      const tabs = this.tabs;
+
+      tabs[activeIndex].setActive(true);
+      if (lastActiveIndex >= 0) {
+        const prevTab = tabs[lastActiveIndex];
+        prevTab.setActive(false);
+        prevTab.pageElement && prevTab.pageElement._hide();
+      }
+    }
+
+    return canceled
+  }
+
+  _onScroll(index, options = {}) {
+    if (this._tabbarBorder) {
+      this._tabbarBorder.style.transition = `all ${options.duration || 0}s ${options.timing || ''}`;
+
+      if (this._autogrow && this._tabsRect.length > 0) {
+        const a = Math.floor(index), b = Math.ceil(index), r = index % 1;
+        this._tabbarBorder.style.width = lerp(this._tabsRect[a].width, this._tabsRect[b].width, r) + 'px';
+        this._tabbarBorder.style.transform = `translate3d(${lerp(this._tabsRect[a].left, this._tabsRect[b].left, r)}px, 0, 0)`;
+      } else {
+        this._tabbarBorder.style.transform = `translate3d(${index * 100}%, 0, 0)`;
+      }
+    }
+
+    this._onSwipe && this._onSwipe(index, options);
+  }
+
+  _onRefresh() {
+    this._autogrow = util.hasModifier(this, 'autogrow');
+    this._tabsRect = this.tabs.map(tab => tab.getBoundingClientRect());
+    if (this._tabbarBorder) {
+      this._tabbarBorder.style.display = this.hasAttribute('tab-border') || platform.isAndroid() ? 'block' : 'none';
+      const index = this.getActiveTabIndex();
+      if (this._tabsRect.length > 0 && index >= 0) {
+        this._tabbarBorder.style.width = this._tabsRect[index].width + 'px';
+      }
+    }
+  }
+
+  _getAutoScrollRatio(matches, velocity, size) {
+    const ratio = .6 + velocity * 1.2 * (matches ? -1 : 1);
+    return Math.min(1, Math.max(0, ratio));
   }
 
   get _tabbarElement() {
     return util.findChild(this, '.tabbar');
   }
 
+  get _contentElement() {
+    return util.findChild(this, '.tabbar__content');
+  }
+
+  get _targetElement() {
+    const content = this._contentElement;
+    return content && content.children[0] || null;
+  }
+
   _compile() {
-    if (this._contentElement && this._tabbarElement) {
-      const content = util.findChild(this, '.tabbar__content');
-      const bar = util.findChild(this, '.tabbar');
+    autoStyle.prepare(this);
 
-      content.classList.add('ons-tabbar__content');
-      bar.classList.add('ons-tabbar__footer');
-    } else {
+    const content = this._contentElement || util.create('.tabbar__content');
+    content.classList.add('ons-tabbar__content');
+    const tabbar = this._tabbarElement || util.create('.tabbar');
+    tabbar.classList.add('ons-tabbar__footer');
 
-      const content = util.create('.ons-tabbar__content.tabbar__content');
-      const tabbar = util.create('.tabbar.ons-tabbar__footer');
-
+    if (!tabbar.parentNode) {
       while (this.firstChild) {
         tabbar.appendChild(this.firstChild);
       }
-
-      this.appendChild(content);
-      this.appendChild(tabbar);
     }
+
+    const activeIndex = Number(this.getAttribute('activeIndex')); // 0 by default
+    if (tabbar.children.length > activeIndex && !util.findChild(tabbar, '[active]')) {
+      tabbar.children[activeIndex].setAttribute('active', '');
+    }
+
+    this._tabbarBorder = util.findChild(tabbar, '.tabbar__border') || util.create('.tabbar__border');
+    tabbar.appendChild(this._tabbarBorder);
+    tabbar.classList.add('ons-swiper-tabbar'); // Hides material border
+
+    !content.children[0] && content.appendChild(document.createElement('div'));
+    !content.children[1] && content.appendChild(document.createElement('div'));
+    content.appendChild = content.appendChild.bind(content.children[0]);
+    content.insertBefore = content.insertBefore.bind(content.children[0]);
+
+    this.appendChild(content);
+    this.appendChild(tabbar); // Triggers ons-tab connectedCallback
+
+    ModifierUtil.initModifier(this, scheme);
   }
 
   _updatePosition(position = this.getAttribute('position')) {
@@ -244,11 +354,13 @@ export default class TabbarElement extends BaseElement {
     const page = util.findParent(this, 'ons-page');
     if (page) {
       contentReady(page, () => {
-        this.style.top = top ? window.getComputedStyle(page._getContentElement(), null).getPropertyValue('padding-top') : '';
-
+        let p = 0;
         if (page.children[0] && util.match(page.children[0], 'ons-toolbar')) {
           action(page.children[0], 'noshadow');
+          p = 1; // Visual fix for some devices
         }
+
+        this.style.top = top ? parseInt(window.getComputedStyle(page._getContentElement(), null).getPropertyValue('padding-top'), 10) - p + 'px' : '';
       });
     }
 
@@ -258,80 +370,20 @@ export default class TabbarElement extends BaseElement {
     });
   }
 
-  _getTabbarElement() {
-    return util.findChild(this, '.tabbar');
-  }
-
-  /**
-   * @return {String}
-   */
-  getTabbarId() {
-    return this._tabbarId;
-  }
-
-  /**
-   * @return {Element/null}
-   */
-  _getCurrentPageElement() {
-    const pages = this._contentElement.children;
-    let page = null;
-    for (var i = 0; i < pages.length; i++) {
-      if (pages[i].style.display !== 'none') {
-        page = pages[i];
-        break;
-      }
-    }
-
-    if (page && page.nodeName.toLowerCase() !== 'ons-page') {
-      throw new Error('Invalid state: page element must be a "ons-page" element.');
-    }
-
-    return page;
-  }
-
   get topPage() {
-    return this._getCurrentPageElement();
+    const tabs = this.tabs,
+      index = this.getActiveTabIndex();
+    return tabs[index]
+      ? tabs[index].pageElement || this.pages[0] || null
+      : null;
   }
 
   get pages() {
-    return util.arrayFrom(this._contentElement.children);
+    return util.arrayFrom(this._targetElement.children);
   }
 
-  /**
-   * @param {Element} element
-   * @param {Object} options
-   * @param {String} [options.animation]
-   * @param {Function} [options.callback]
-   * @param {Object} [options.animationOptions]
-   * @param {Number} options.selectedTabIndex
-   * @param {Number} options.previousTabIndex
-   * @return {Promise} Resolves to the new page element.
-   */
-  _switchPage(element, options) {
-    const oldPageElement = this._oldPageElement || internal.nullElement;
-    this._oldPageElement = element;
-    const animator = this._animatorFactory.newAnimator(options);
-
-    return new Promise(resolve => {
-      if (oldPageElement !== internal.nullElement) {
-        oldPageElement._hide();
-      }
-
-      animator.apply(element, oldPageElement, options.selectedTabIndex, options.previousTabIndex, () => {
-        if (oldPageElement !== internal.nullElement) {
-          oldPageElement.style.display = 'none';
-        }
-
-        element.style.display = 'block';
-        element._show();
-
-        if (options.callback instanceof Function) {
-          options.callback();
-        }
-
-        resolve(element);
-      });
-    });
+  get tabs() {
+    return Array.prototype.filter.call(this._tabbarElement.children, e => e.tagName === 'ONS-TAB');
   }
 
   /**
@@ -343,12 +395,12 @@ export default class TabbarElement extends BaseElement {
    * @param {Object} [options]
    *   [en]Parameter object.[/en]
    *   [ja]オプションを指定するオブジェクト。[/ja]
-   * @param {Boolean} [options.keepPage]
-   *   [en]If true the page will not be changed.[/en]
-   *   [ja]タブバーが現在表示しているpageを変えない場合にはtrueを指定します。[/ja]
+   * @param {Function} [options.callback]
+   *   [en]Function that runs when the new page has loaded.[/en]
+   *   [ja][/ja]
    * @param {String} [options.animation]
-   *   [en]Animation name. Available animations are `"fade"`, `"slide"` and `"none"`.[/en]
-   *   [ja]アニメーション名を指定します。`"fade"`、`"slide"`、`"none"`のいずれかを指定できます。[/ja]
+   *   [en]If this option is "none", the transition won't slide.[/en]
+   *   [ja][/ja]
    * @param {String} [options.animationOptions]
    *   [en]Specify the animation's duration, delay and timing. E.g. `{duration: 0.2, delay: 0.4, timing: 'ease-in'}`.[/en]
    *   [ja]アニメーション時のduration, delay, timingを指定します。e.g. {duration: 0.2, delay: 0.4, timing: 'ease-in'}[/ja]
@@ -356,87 +408,39 @@ export default class TabbarElement extends BaseElement {
    *   [en]Show specified tab page. Animations and other options can be specified by the second parameter.[/en]
    *   [ja]指定したインデックスのタブを表示します。アニメーションなどのオプションを指定できます。[/ja]
    * @return {Promise}
-   *   [en]Resolves to the new page element.[/en]
+   *   [en]A promise that resolves to the new page element.[/en]
    *   [ja][/ja]
    */
-  setActiveTab(index, options = {}) {
-    if (options && typeof options != 'object') {
-      throw new Error('options must be an object. You supplied ' + options);
-    }
+  setActiveTab(nextIndex, options = {}) {
+    const prevIndex = this.getActiveTabIndex();
+    const prevTab = this.tabs[prevIndex],
+      nextTab = this.tabs[nextIndex];
 
-    options.animationOptions = util.extend(
-      options.animationOptions || {},
-      AnimatorFactory.parseAnimationOptionsString(this.getAttribute('animation-options'))
-    );
-
-    if (!options.animation && this.hasAttribute('animation')) {
-      options.animation = this.getAttribute('animation');
-    }
-
-    const previousTab = this._getActiveTabElement(),
-      selectedTab = this._getTabElement(index),
-      previousTabIndex = this.getActiveTabIndex(),
-      selectedTabIndex = index,
-      previousPageElement = this._getCurrentPageElement();
-
-    if (!selectedTab) {
+    if (!nextTab) {
       return Promise.reject('Specified index does not match any tab.');
     }
 
-    if (selectedTabIndex === previousTabIndex) {
-      util.triggerElementEvent(this, 'reactive', {
-        index: selectedTabIndex,
-        tabItem: selectedTab
-      });
-
-      return Promise.resolve(previousPageElement);
+    if (nextIndex === prevIndex) {
+      util.triggerElementEvent(this, 'reactive', { index: nextIndex, activeIndex: nextIndex, tabItem: nextTab });
+      return Promise.resolve(nextTab.pageElement);
     }
 
-    let canceled = false;
-
-    util.triggerElementEvent(this, 'prechange', {
-      index: selectedTabIndex,
-      tabItem: selectedTab,
-      cancel: () => canceled = true
-    });
-
-    if (canceled) {
-      selectedTab.setInactive();
-      if (previousTab) {
-        previousTab.setActive();
-      }
-      return Promise.reject('Canceled in prechange event.');
-    }
-
-    selectedTab.setActive();
-
-    const params = {
-      ...options,
-      previousTabIndex: previousTabIndex,
-      selectedTabIndex: selectedTabIndex
-    };
-
-    if (previousTab) {
-      previousTab.setInactive();
-    } else {
-      params.animation = 'none';
-    }
-
-    return new Promise(resolve => {
-      selectedTab._loadPageElement(this._contentElement, pageElement => {
-        pageElement.removeAttribute('style');
-
-        this._switchPage(pageElement, params)
-          .then(page => {
-            util.triggerElementEvent(this, 'postchange', {
-              index: selectedTabIndex,
-              tabItem: selectedTab
-            });
-
-            return resolve(page);
-          });
-      });
-    });
+    // FIXME: nextTab.loaded is broken in Zone.js promises (Angular2)
+    const nextPage = nextTab.pageElement;
+    return (nextPage ? Promise.resolve(nextPage) : nextTab.loaded)
+      .then(nextPage => this._swiper.setActiveIndex(nextIndex, {
+        reject: true,
+        ...options,
+        animation: prevTab && nextPage ? options.animation || this.getAttribute('animation') : 'none',
+        animationOptions: util.extend(
+          { duration: .3, timing: 'cubic-bezier(.4, .7, .5, 1)' },
+          this.hasAttribute('animation-options') ? util.animationOptionsParse(this.getAttribute('animation-options')) : {},
+          options.animationOptions || {}
+        )
+      }).then(() => {
+        options.callback instanceof Function && options.callback(nextPage);
+        return nextPage;
+      }));
   }
 
   /**
@@ -449,7 +453,7 @@ export default class TabbarElement extends BaseElement {
    */
   setTabbarVisibility(visible) {
     this._contentElement.style[this._top ? 'top' : 'bottom'] = visible ? '' : '0px';
-    this._getTabbarElement().style.display = visible ? '' : 'none';
+    this._tabbarElement.style.display = visible ? '' : 'none';
   }
 
   show() {
@@ -469,7 +473,40 @@ export default class TabbarElement extends BaseElement {
    *   [ja]タブバーが見える場合に`true`。[/ja]
    */
   get visible() {
-    return this._getTabbarElement().style.display !== 'none';
+    return this._tabbarElement.style.display !== 'none';
+  }
+
+  /**
+   * @property swipeable
+   * @type {Boolean}
+   * @description
+   *   [en]Enable swipe interaction.[/en]
+   *   [ja]swipeableであればtrueを返します。[/ja]
+   */
+  get swipeable() {
+    return this.hasAttribute('swipeable');
+  }
+
+  set swipeable(value) {
+    return util.toggleAttribute(this, 'swipeable', value);
+  }
+
+  /**
+   * @property onSwipe
+   * @type {Function}
+   * @description
+   *   [en]Hook called whenever the user slides the tabbar. It gets a decimal index and an animationOptions object as arguments.[/en]
+   *   [ja][/ja]
+   */
+  get onSwipe() {
+    return this._onSwipe;
+  }
+
+  set onSwipe(value) {
+    if (value && !(value instanceof Function)) {
+      throw new Error(`'onSwipe' must be a function.`)
+    }
+    this._onSwipe = value;
   }
 
   /**
@@ -483,57 +520,30 @@ export default class TabbarElement extends BaseElement {
    *   [ja]現在アクティブになっているタブのインデックスを返します。現在アクティブなタブがない場合には-1を返します。[/ja]
    */
   getActiveTabIndex() {
-    const tabs = this._getTabbarElement().children;
-
-    for (var i = 0; i < tabs.length; i++) {
-      if (tabs[i] instanceof TabElement && tabs[i].isActive && tabs[i].isActive()) {
+    for (let tabs = this.tabs, i = 0; i < tabs.length; i++) {
+      if (tabs[i] instanceof TabElement && tabs[i].isActive()) {
         return i;
       }
     }
-
     return -1;
   }
 
-  /**
-   * @return {Number} When active tab is not found, returns -1.
-   */
-  _getActiveTabElement() {
-    return this._getTabElement(this.getActiveTabIndex());
-  }
-
-  /**
-   * @return {Element}
-   */
-  _getTabElement(index) {
-    return this._getTabbarElement().children[index];
-  }
-
-  disconnectedCallback() { }
-
   _show() {
-    const currentPageElement = this._getCurrentPageElement();
-    if (currentPageElement) {
-      currentPageElement._show();
-    }
+    setImmediate(() => this.tabs[this.getActiveTabIndex()].loaded.then(el => el && el._show()))
   }
 
   _hide() {
-    const currentPageElement = this._getCurrentPageElement();
-    if (currentPageElement) {
-      currentPageElement._hide();
-    }
+    const topPage = this.topPage;
+    topPage && topPage._hide();
   }
 
   _destroy() {
-    const tabs = this._getTabbarElement().children;
-    for (let i = tabs.length - 1; i >= 0; i--) {
-      tabs[i].remove();
-    }
+    this.tabs.forEach(tab => tab.remove());
     this.remove();
   }
 
   static get observedAttributes() {
-    return ['modifier', 'position'];
+    return ['modifier', 'position', 'swipeable', 'tab-border'];
   }
 
   attributeChangedCallback(name, last, current) {
@@ -541,8 +551,10 @@ export default class TabbarElement extends BaseElement {
       ModifierUtil.onModifierChanged(last, current, this, scheme);
       const isTop = m => /(^|\s+)top($|\s+)/i.test(m);
       isTop(last) !== isTop(current) && this._updatePosition();
-    } else if(name === 'position') {
+    } else if (name === 'position') {
       this._updatePosition();
+    } else if (name === 'swipeable') {
+      this._swiper && this._swiper.updateSwipeable(this.hasAttribute('swipeable'));
     }
   }
 
@@ -550,27 +562,8 @@ export default class TabbarElement extends BaseElement {
     return rewritables;
   }
 
-  static get TabbarAnimator() {
-    return TabbarAnimator;
-  }
-
   static get events() {
     return ['prechange', 'postchange', 'reactive'];
-  }
-
-  /**
-   * @param {String} name
-   * @param {Function} Animator
-   */
-  static registerAnimator(name, Animator) {
-    if (!(Animator.prototype instanceof TabbarAnimator)) {
-      throw new Error('"Animator" param must inherit TabbarElement.TabbarAnimator');
-    }
-    _animatorDict[name] = Animator;
-  }
-
-  static get animators() {
-    return _animatorDict;
   }
 }